vcf: expose genotypes from VCF
as of this commit, the read functionality for VCF is relatively
complete. The access to genotypes is powerful, convenient, and
efficient. They are left as int32's but given methods that make
them appear as objects to the user. In all cases, the user is
given the oportunity to re-use memory.

Much of the inpsiration for this way of doing it it taken from
@johanneskoester 's excellent implementation in rust-htslib with
changes here to make it more nim-like.